Backend/requirements.txt
fastapi
uvicorn
langchain
langchain-google-genai
langgraph
python-dotenv
pymongo
httpx
beautifulsoup4
pydantic
tavily-python
python-jose[cryptography]
passlib[bcrypt]
python-multipart
motor
email-validator
reportlab
pinecone-client
langchain-pinecone
pypdf
requests
langchain-community
langchain-text-splitters
langchain-xai
transformers
torch
langchain-huggingface
bcrypt
langchain-core
google-generativeai
accelerate
accelerate
peft
safetensors
sentence-transformers
